+++ src/swganh/messages/client_permissions_message.h
@@ -18,18 +18,17 @@
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
-#ifndef SWGANH_CONNECTION_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
-#define SWGANH_CONNECTION_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
+#ifndef SWGANH_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
+#define SWGANH_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
 
 #include <cstdint>
 #include "anh/byte_buffer.h"
-#include "swganh/base/swg_message.h"
+#include "swganh/messages/base_swg_message.h"
 
 namespace swganh {
-namespace connection {
 namespace messages {
     
-struct ClientPermissionsMessage : public swganh::base::SwgMessage<ClientPermissionsMessage> {
+struct ClientPermissionsMessage : public swganh::messages::BaseSwgMessage<ClientPermissionsMessage> {
     static uint16_t opcount() { return 4; }
     static uint32_t opcode() { return 0xE00730E5; }
     
@@ -50,6 +49,6 @@ struct ClientPermissionsMessage : public swganh::base::SwgMessage<ClientPermissi
     }
 };
 
-}}}  // namespace swganh::connection::messages
+}}  // namespace swganh::messages
 
-#endif  // SWGANH_CONNECTION_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
+#endif  // SWGANH_MESSAGES_CLIENT_PERMISSIONS_MESSAGE_H_
